added the next failing test - this considers the requirement where we need to show the total value of the bitcoin portfolio

+++ src/main/java/portfolio/BitcoinPortfolio.java
@@ -16,7 +16,7 @@ public class BitcoinPortfolio {
         this.fileApi = fileApi;
     }
 
-    public double calculatePortfolioValue(Map<String, Double> totalValuePerBitcoin) {
+    public double calculatePortfolioValue(Map<String, BigDecimal> totalValuePerBitcoin) {
         return 0.0;
     }
